rename:이미지 업로더 파일 이름 수정"
